requirements.txt
celery
Django
django-allauth
django-cuser
django-filter
django-modeltranslation
django-property-filter
django-rest-passwordreset
django-tinymce
djangorestframework
psycopg2-binary
requests
virtualenv
virtualenv-clone
ujson
pyyaml
redis
pytest
pytest-django
model_bakery
drf_spectacular
drf_spectacular_sidecar
social-auth-app-django